Integrate chai-subset into chai
