+++ lib/chai/core/assertions.js
@@ -8,6 +8,7 @@
 import {Assertion} from '../assertion.js';
 import {AssertionError} from 'assertion-error';
 import * as _ from '../utils/index.js';
+import {config} from '../config.js';
 
 const {flag} = _;
 
@@ -3929,5 +3930,65 @@ Assertion.addProperty('finite', function(msg) {
       typeof obj === 'number' && isFinite(obj)
     , 'expected #{this} to be a finite number'
     , 'expected #{this} to not be a finite number'
+});
+
+function compare(expected, actual) {
+  if (expected === actual) {
+    return true;
+  }
+  if (typeof actual !== typeof expected) {
+    return false;
+  }
+  if (typeof expected !== 'object' || expected === null) {
+    return expected === actual;
+  }
+  if (!!expected && !actual) {
+    return false;
+  }
+
+  if (Array.isArray(expected)) {
+    if (typeof actual.length !== 'number') {
+      return false;
+    }
+    var aa = Array.prototype.slice.call(actual);
+    return expected.every(function (exp) {
+      return aa.some(function (act) {
+        return compare(exp, act);
+      });
+    });
+  }
+
+  if (expected instanceof Date) {
+    if (actual instanceof Date) {
+      return expected.getTime() === actual.getTime();
+    } else {
+      return false;
+    }
+  }
+
+  return Object.keys(expected).every(function (key) {
+    var eo = expected[key];
+    var ao = actual[key];
+    if (typeof eo === 'object' && eo !== null && ao !== null) {
+      return compare(eo, ao);
+    }
+    if (typeof eo === 'function') {
+      return eo(ao);
+    }
+    return ao === eo;
+  });
+}
+
+Assertion.addMethod('containSubset', function (expected) {
+  const actual = _.flag(this, 'object');
+  const showDiff = config.showDiff;
+
+  this.assert(
+    compare(expected, actual),
+    'expected #{act} to contain subset #{exp}',
+    'expected #{act} to not contain subset #{exp}',
+    expected,
+    actual,
+    showDiff
   );
 });
