+++ lib/chai/interface/assert.js
@@ -2985,6 +2985,10 @@ assert.isNotEmpty = function(val, msg) {
   new Assertion(val, msg, assert.isNotEmpty, true).to.not.be.empty;
 };
 
+assert.containSubset = function (val, exp, msg) {
+  new Assertion(val, msg).to.be.containSubset(exp);
+};
+
 /**
  * Aliases.
  *
